api/requirements.txt
fastapi==0.115.10
uvicorn[standard]
pika==1.3.2
python-dotenv==1.0.1
mysql-connector-python==9.2.0
